Finished 1&2
First, let’s take a look at our index.html file. As you can see, we have several <div>s that contain the content of our website.

Find the <div> with the class grid — this is our container grid and the properties applied to it will determine the structure for the content inside.

Next, locate the <div>s that share the class box and the <div>s that share the class testimonial. Their positioning will change the most when the new grid properties are added.

You can check this step off when you feel like you’re ready to proceed!

2.
Now let’s look at our CSS. We have provided the initial styling. Now it is up to you to add the necessary CSS grid properties to make sure the content is properly laid out on the page. The first step will be to initialize the grid, and then we’re going to specify the number and size of our rows and columns.

In style.css, inside the .grid ruleset, set the display property to grid.