Finished 6
Now, let’s fix that broken page layout. The content in the following rulesets needs to extend across all six columns:

header
.banner
.about
footer
Add the grid-column-start and grid-column-end properties to these rulesets, with values that start the content at the the first column and end it at the sixth column.