Finished 1
We will be focusing on laying out the container <div> and the two card-column <div>s within it. Here’s an abbreviated version of those <div>s in index.html:

<div class="container">
    <div class="card-column future-projects">
        ...
    </div>
    <div class="card-column active-projects">
        ...
    </div>
</div>

Right now, the task board looks pretty cluttered. In style.css, let’s make the <div> with class container a grid container so that we can start to arrange these tasks better.