Finished 4
Inside of each card column, there are 5 rows of task cards. We want them to be arranged neatly within the columns as well, so let’s make each card column a grid container.

In style.css, within the .card-column ruleset, add a declaration that sets the display property to grid.