Finished 7 (soo easy)
Now we want to add another card column for Completed Projects to keep track of all of the tasks we’ve already done. Let’s add a third card column that’s like the other two. We already have our container set to have space for three columns, so you just need to add the HTML.

In index.html, uncomment the code for the third card column.