Finished 8 & 9
8.
We expect to complete a lot of projects, but we don’t know how many. Since those projects will be added as task cards to the Completed Projects column, it will be helpful to use implicit grid properties instead of the explicit properties we have been using so far. We know the first row with the heading should be explicitly 20px, just as the other heading rows are. Every additional row, if it exists, should be implicitly set.

Inside .card-column, use the grid-auto-rows property to implicitly set rows to 100px.
9.
Good job! Now, feel free to play around with the layout of this site, using the other CSS properties you’ve learned. Once you’re satisfied with your work, press the Next Button to move on.